chore: Update npm dependencies and plugins
Update the npm dependencies and plugins in the package.json file. This includes updating the version of "@faker-js/faker" to "^9.0.2" and "react-router-dom" to "^6.26.2". Also update the devDependencies, including "@types/react" to "^18.3.8", "@typescript-eslint/eslint-plugin" to "^8.7.0", and "eslint-plugin-react-refresh" to "^0.4.12". Finally, update the versions of "msw" to "2.4.9", "postcss" to "^8.4.47", "tailwindcss" to "^3.4.13", "typescript" to "^5.6.2", and "vite" to "^5.4.7".